requirements.txt
fastapi >=0.108.0
uvicorn >=0.25.0
tortoise-orm>=0.20.0